[refactor] #133 Category Live, Test 분리
